requirements.txt
Django>=5.1,<6.0
django-environ>=0.11,<1.0
psycopg[binary]>=3.2,<4.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0

celery>=5.4,<6.0
redis>=5.0,<6.0

openai>=1.30,<2.0
stripe>=10.0,<11.0

pypdf>=4.2,<5.0

# Vector support
pgvector>=0.3,<1.0

# Optional but useful
uvicorn>=0.30,<1.0
